requirements.txt
uvicorn==0.17.0
python-decouple==3.5
asgiref==3.4.1
Flask==2.1.2
loguru~=0.6.0
aioflask~=0.4.0
python-dotenv==0.20.0
dnspython==2.2.1
pymongo==4.2.0